Add files via upload
his website gives you a classic method for composing secret messages, which is called a square code. The spaces are removed from the english text and the characters are written into a square (or rectangle). The user can enter a message in with no spaces between the words. Maximum message length to be 81 characters. A warning will be displayed if there is some wrong input.
Example:
anabajric
ana baj ric
abr nai ajc<br>
The project consists of forms and javascript logic used to handle the operations.